day05 ex01

+++ day05/ex01/myUselessChat/src/main/java/edu/school21/chat/repositories/MessagesRepository.java
@@ -6,5 +6,5 @@ import java.sql.SQLException;
 import java.util.Optional;
 
 public interface MessagesRepository {
-    Optional<Message> findById(Long id) throws SQLException;
+    Optional<Message> findById(int id) throws SQLException;
 }
